% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BubbleTree.R
\name{adjust_plot}
\alias{adjust_plot}
\alias{drawBubblePlot}
\title{Draw a BubblePlot from a list of pre-processed segments}
\usage{
adjust_plot()
}
\value{
Plots a full BubblePlot using pre-processed segment data
}
\description{
Draw a Bubble plot using pre-processed segment data that
contains the following:

seg.mean - copy ratio score of the segment
hds.median - median HDS score of the segment
num.mark - number of the marks harbored by the segment
col - color of the bubble
min.cex - minimum font size
size - size of the bubble to scale
info - label of the bubble
adj - adjusted postion of the label
}

