% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{btcompare}
\alias{btcompare}
\title{btcompare}
\usage{
btcompare(set1, set2)
}
\arguments{
\item{set1}{first set}

\item{set2}{second set to compare}
}
\value{
combined, unique list of genes
}
\description{
btcompare
}
\examples{

load(system.file("data", "cancer.genes.minus2.rda", package="BubbleTree"))
load(system.file("data", "vol.genes.rda", package="BubbleTree"))

# 77 common cancer genes
comm <- btcompare(vol.genes, cancer.genes.minus2)
}
