% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BTreePredictor.R
\docType{package}
\name{BTreePredictor}
\alias{BTreePredictor}
\alias{BTreePredictor-package}
\title{BTreePredictor}
\description{
BTreePredictor
}
\examples{
btreepredictor <- new("BTreePredictor")
}

