% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbletree.R
\name{bubbletree}
\alias{bubbletree}
\title{<Add Title>}
\usage{
bubbletree(
  tree_data,
  width = NULL,
  height = NULL,
  elementId = NULL,
  settings = list(smoothness = 30, padding = 10, resizing_factor = 20, max_textsize =
    0.3, min_textsize_scale = 0.1)
)
}
\description{
<Add Description>
}
