% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{calc.prev}
\alias{calc.prev}
\title{Calculate tumor cell prevalence in a sample, an indication of sample purity}
\usage{
calc.prev(rbdx,heurx=FALSE,modex=3,plotx="~/prev_model.pdf")
}
\arguments{
\item{rbdx}{a matrix or data frame generated by the function plotBubble().}

\item{heurx}{a logical value indicating if only using A/B and AA/BB
branches for purity calculation or all branches.}

\item{modex}{an integer value providing the expected number of modes in
mixture distribution.}

\item{plotx}{a character string specifying the mixture model plot file
name.}
}
\value{
List object of two elements: 1) the rbdx data frame with two
addition columns (2^seg.mean and genotype with frequency for each segment),
 and 2) a data matrix of modex rows and two columns indicating the seeding
  modes (column 1) and estimated modes (column 2), each with the number of
   segments supporting each mode, separated by an underscore. The largest
    mode in column two is the estimated tumor purity
}
\description{
A model fitting of component distributions calculated from
somatic copy number segments within the BubbleTree diagram.
}
\details{
The top n (user defined) most frequent prevalence estimates are
seeded as means in this finite mixture model, to predict the tumor
(sub)clonal prevalences. This allows a user-defined expected number of
modes within the component distribution, though overlapping modes converge
to the same estimate. Then these SCNA segment frequencies are used to
estimate the means (+/- standard deviations) of the component distributions
using the R package mixdist [Macdonald et al, 1988]. The standard deviations
are constrained by the Poisson relation given by
 \eqn{\alpha=\sqrt{u_i} ,i=1,...,k}{\alpha=\sqrt(u_i) ,i=1,...,k}.  .
}
\examples{
#load sequence variants
data('hetero.gr', package='BubbleTree', envir = environment())
#load copy number variation data
data('cnv.gr', package='BubbleTree', envir = environment())
rbd<-getRBD(hetero.gr, cnv.gr) #plot BubbleTree
pur <- calc.prev(rbdx=rbd,heurx=FALSE,modex=3,plotx="prev_model.pdf")

# extract the genotype (branch) and frequency for each segment
 pur[[1]]$ploidy_prev
 pur[[2]][nrow(pur[[2]]),2]
}
\references{
Macdonald PDM. and Green PEJ: User's Guide to Program MIX: An Interactive
 Program for Fitting Mixtures of Distributions. ICHTHUS DATA SYSTEMS 1988.
Macdonald, PDM (1988). \emph{Demonstration Examples for MIX 2.3}. Ichthus
 Data Systems, Hamilton, Ontario. 13 pp. ISBN 0-9692305-1-4.
}

