% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\docType{data}
\name{cnv.gr}
\alias{cnv.gr}
\title{A sample dataset of tumor segmented copy number variations}
\format{The format is: Formal class 'GRanges' [package "GenomicRanges"] with 6 slots}
\usage{
data(cnv.gr)
}
\description{
A GRanges object containing segmented copy number log2 ratios
            and number of markers/segemnt in a tumor/normal sample pair from a patient
            with NSCLC
}
\details{
Metadata columns include
\itemize{
 \item \strong{num.mark} number of markers within segment
 \item \strong{seg.mean} mean log2 copy number ratio within a segment
 }
}
\examples{
data(cnv.gr)
}
\keyword{datasets}

