% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{compareBubbles}
\alias{compareBubbles}
\title{compare bubbles from two samples}
\usage{
compareBubbles(rbd1, rbd2, min.mark=500, min.dist=0.2, max.dist=100, main="")
}
\arguments{
\item{rbd1}{RBD (R-score BAF Dataframe) from the sample 1}

\item{rbd2}{RBD data.frame from the sample 2}

\item{min.mark}{integer segments with mininum markers to be compared}

\item{min.dist}{numeric minimum distance of the overlapped segments to be displayed}

\item{max.dist}{numeric maximum distance of the overlapped segments to be displayed}

\item{main}{character string for the plot title}
}
\value{
A list of the detailed information of the overlapped segments
}
\description{
TBA
}
\details{
The segments (larger than min.mark) from the two samples are compared
 to each other.
}
\examples{
data('hcc.rbd.lst', package='BubbleTree', envir = environment())

# show the SCNV changes between the recurrent tumor and the primary tumor
compareBubbles(hcc.rbd.lst$HCC11.Primary.Tumor,
    hcc.rbd.lst$HCC11.Recurrent.Tumor,min.dist=0.05, min.mark=2000)

# show the similarity in the recurrent tumors between two subjects
# Interestingly, 17p- and 17q+ are conserved.
compareBubbles(hcc.rbd.lst$HCC4.Recurrent.Tumor,
    hcc.rbd.lst$HCC11.Recurrent.Tumor,
    min.dist=0.0, max.dist=0.1, min.mark=500)
}

