% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{drawBranches}
\alias{drawBranches}
\title{Plot branches of BubbleTree plot}
\usage{
drawBranches(xmax=3.2, main="")
}
\arguments{
\item{xmax}{define the upper limit of the x-axis}

\item{main}{title of the plot}
}
\value{
A plot showing branches of a BubbleTree
}
\description{
Plot branches of BubbleTree plot
}
\details{
The branches of BubbleTree plot stand for interger copy number change.
  For example, "B" and "BB" indicates LOH and copy-number neutral LOH, repectively.
}
\examples{
drawBranches(xmax=2.6)
drawBranches()
}

