% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{drawBubble}
\alias{addBubble}
\alias{drawBubble}
\title{Draw a bubble}
\usage{
drawBubble(seg.mean, hds.median, num.mark, col, min.cex=0.3,
  size=1, info=NULL ,adj=0.5)
}
\arguments{
\item{seg.mean}{copy ratio score of the segment}

\item{hds.median}{median HDS score of the segment}

\item{num.mark}{number of the marks harbored by the segment}

\item{col}{color of the bubble}

\item{min.cex}{minimum font size}

\item{size}{size of the bubble to scale}

\item{info}{label of the bubble}

\item{adj}{adjusted postion of the label}
}
\value{
Plots a single bubble on the BubbleTree Plot
}
\description{
Draw single bubble to BubbleTree plot with customized
 label, size and color
}
\examples{
drawBranches(main="Demo")
drawBubble(0.5, 0.3, 5000, col="blue", size=2, info="PTEN", adj=-0.5)
}

