% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTreePlotter.R
\docType{methods}
\name{drawBubbles}
\alias{drawBubbles}
\alias{drawBubbles,BTreePlotter-method}
\title{drawBubbles}
\usage{
drawBubbles(.Object, rbd, col = NULL)

\S4method{drawBubbles}{BTreePlotter}(.Object, rbd, col = "gray80")
}
\arguments{
\item{.Object}{the object}

\item{rbd}{the rbd object}

\item{col}{the col value}
}
\value{
draw the bubbles on the track
}
\description{
draw the Bubbles
}
\examples{

load(system.file("data", "allCall.lst.RData", package="BubbleTree"))

btreeplotter <- new("BTreePlotter", max.ploidy=5, max.size=10)
nn <- "sam2"
rbd1 <- allCall.lst[[nn]]@rbd
rbd2 <- allCall.lst[[nn]]@rbd.adj
arrows <- trackBTree(btreeplotter, rbd1, rbd2, min.srcSize=0.01, 
                     min.trtSize=0.01)
btree <- drawBTree(btreeplotter, rbd1) + 
    drawBubbles(btreeplotter, rbd2, "gray80") + arrows
}

