% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{getRBD}
\alias{getRBD}
\title{Get RBD (R-score BAF Dataframe) of the homogeneous SCNV segments}
\usage{
getRBD(snp.gr, cnv.gr, max.sd = 0.1)
}
\arguments{
\item{snp.gr}{a GRanges object containing BAF (B-allele frequency) of the germline
heterogenous loci}

\item{cnv.gr}{a GRanges object containing num.mark and seg.mean, generated
from the CNV call}

\item{max.sd}{Numeric value indicating the maximum standard
deviation of Homozygous Deviation Scores (HDS) within a cnv segment.
Segments with SD above this cutoff will be omitted.}
}
\value{
A data frame to be called by plotBubble
}
\description{
Get RBD (R-score BAF Dataframe) of the homogeneous SCNV segments
}
\details{
This function merge BAF and CNV call results into one data frame.
 The segments with high
HDS variation are omitted. The RBD of the remaining "homogeneous" segments are
 returned.
}
\examples{
#load sequence variants
data('hetero.gr', package='BubbleTree', envir = environment())
#load copy number variation data
data('cnv.gr', package='BubbleTree', envir = environment())
rbd <- getRBD(snp.gr=hetero.gr, cnv.gr=cnv.gr)
plotBubbles(rbd, main="BubbleTree Plot")
}

