% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\docType{data}
\name{hcc.rbd.lst}
\alias{hcc.rbd.lst}
\title{Rbd data from the HCC samples}
\format{The format is: A list of 4 data frames}
\usage{
data(hcc.rbd.lst)
}
\description{
A list of RBD (R-BAF Dataframe) of the four HCC samples:\cr
  HCC4.Primary.Tumor \cr
  HCC4.Recurrent.Tumor\cr
  HCC11.Primary.Tumor\cr
  HCC11.Recurrent.Tumor\cr
}
\details{
These data frames are produced by the getRBD() function and include columns for:
\itemize{
  \item \strong{seg.id} copy number segment identifier
  \item \strong{hds.median}  median homozygosity deviation score withing the segment
  \item \strong{hds.sd} standard deviation of hds within the segment
  \item \strong{num.mark} number of markers within the segment
  \item \strong{seg.mean} mean copy number of markers within segment
  \item \strong{chr} segment chromosome
  \item \strong{start} segment start coordinate
  \item \strong{end} segment end coordinate
  \item \strong{cytoband} segment cytoband
 }
}
\examples{
data(hcc.rbd.lst)
}
\keyword{datasets}

