% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\docType{data}
\name{hetero.gr}
\alias{hetero.gr}
\title{BAF of the germline heterozygous loci in GRanges format}
\format{The format is: Formal class 'GRanges' [package "GenomicRanges"] with 6 slots}
\usage{
data(hetero.gr)
}
\description{
A sample data of B-allele frquencies of the germline heterozygous loci
}
\details{
Metadata columns include
\itemize{
 \item \strong{freq} B allele frequency
 }
}
\examples{
data(hetero.gr)
}
\keyword{datasets}

