% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBD.R
\docType{methods}
\name{makeRBD}
\alias{makeRBD}
\alias{makeRBD,RBD-method}
\title{makeRBD}
\usage{
makeRBD(.Object, snp.gr = NULL, cnv.gr = NULL, unimodal.kurtosis = -0.1)

\S4method{makeRBD}{RBD}(.Object, snp.gr = snp.gr, cnv.gr = cnv.gr,
  unimodal.kurtosis = -0.1)
}
\arguments{
\item{snp.gr}{SNP GenomicRanges object}

\item{cnv.gr}{CNV GenomicRanges object}
}
\value{
RBD object
}
\description{
make the RBD object
}
\examples{
load(system.file("extdata", "cnv.test.gr", package="BubbleTree"))
load(system.file("extdata", "snv.test.gr", package="BubbleTree"))

# initialize RBD object
r <- new("RBD", unimodal.kurtosis=-0.1)

# create new RBD object with GenomicRanges objects for SNPs and CNVs
rbd <- makeRBD(r, snp.gr, cnv.gr)
}

