% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BubbleTree.R
\name{medi_compare}
\alias{compareSamples}
\alias{medi_compare}
\title{Compare BubblePlots from 2 different sets of cancer data}
\usage{
medi_compare()
}
\value{
Comparison BubblePlots
}
\description{
Compare BubblePlots from 2 different sets of cancer data
}

