% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BubbleTree.R
\name{plot_tracks}
\alias{plotAllTracks}
\alias{plot_tracks}
\title{A testing function that draws a BubblePlot from a list of pre-processed
segments with overlays of different types of annotation data.}
\usage{
plot_tracks()
}
\value{
A BubblePlot with called CNVs with annotation overlayed
}
\description{
A testing function that draws a BubblePlot from a list of
pre-processed segments with overlays of different types of annotation data.

seg.mean - copy ratio score of the segment
hds.median - median HDS score of the segment
num.mark - number of the marks harbored by the segment
col - color of the bubble
min.cex - minimum font size
size - size of the bubble to scale
info - label of the bubble
adj - adjusted postion of the label
}

