% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbletree_functions.R
\name{plotBubbles}
\alias{plotBubbles}
\title{Plot Bubbles}
\usage{
plotBubbles(rbd,  min.cex=0.3, show.cyto=TRUE, no.bayes=FALSE,
 xmax=3.2, size=1, main="BubbleTree Plot")
}
\arguments{
\item{rbd}{a data.frame containing tumor allele frequency
and segmented CNV previously generated by the getRBD() function}

\item{min.cex}{minimum size of bubble annotation on the plot}

\item{show.cyto}{Logical; indicating if cytoband information
should be displayed on plot. Default is TRUE.}

\item{no.bayes}{Logical: contol labels}

\item{xmax}{maximum value for R score plotted on x-axis}

\item{size}{scaling factor that controls relative size of
bubbles appearing on plot}

\item{main}{character string for the plot title}
}
\value{
Creates a bubbletree plot using the current graphics device and
returns a data.frame object containing summary information on genomic
regions affected by chromosome loss or gain used in the plot
}
\description{
Plot Bubbles
}
\details{
For each segment iteratively calculate the median and
 standard deviation (SD) of the homozygous deviation score (HDS) of the
  heterozygous-loci and filter those SCNAs with
high HDS variation (empirically, SD>0.2). The median of HDS,
the copy ratio and segment size for each homogenous SCNA are used to define
 the X-Y coordinates and sizes of the bubbles in the diagram.
}
\examples{
#load sequence variants
data('hetero.gr', package='BubbleTree', envir = environment())
#load copy number variation data
data('cnv.gr', package='BubbleTree', envir = environment())
rbd <- getRBD(snp.gr=hetero.gr, cnv.gr=cnv.gr)
plotBubbles(rbd)

data('hcc.rbd.lst', package='BubbleTree', envir = environment())
pdf(file="hcc.bubbletree.pdf", width=8, height=6)
lapply(names(hcc.rbd.lst), function(sample) plotBubbles(hcc.rbd.lst[[sample]],
 size=2, main=sample))
dev.off()
}

