% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BubbleTree.R
\name{show_summary}
\alias{showSummaryData}
\alias{show_summary}
\title{Create a summary Excel file}
\usage{
show_summary()
}
\value{
Excel file with summary data
}
\description{
Create a summary Excel file
}

