% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBD.R
\docType{methods}
\name{makeRBD}
\alias{makeRBD}
\alias{makeRBD,RBD-method}
\title{makeRBD}
\usage{
makeRBD(.Object, unimodal.kurtosis = -0.1, snp = NULL, cnv = NULL, ...)

\S4method{makeRBD}{RBD}(.Object, unimodal.kurtosis = -0.1, snp.gr = NULL,
  cnv.gr = NULL)
}
\arguments{
\item{.Object}{the object}

\item{unimodal.kurtosis}{kurtosis}

\item{snp.gr}{SNP GenomicRanges object}

\item{cnv.gr}{CNV GenomicRanges object}
}
\value{
RBD object
}
\description{
make the RBD object
}
\examples{

load(system.file("data", "cnv.gr.rda", package="BubbleTree"))
load(system.file("data", "snp.gr.rda", package="BubbleTree"))

# initialize RBD object
r <- new("RBD", unimodal.kurtosis=-0.1)

# create new RBD object with GenomicRanges objects for SNPs and CNVs
rbd <- makeRBD(r, snp.gr, cnv.gr)
}

