% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBD.R
\docType{methods}
\name{mergeSnpCnv}
\alias{mergeSnpCnv}
\alias{mergeSnpCnv,RBD-method}
\title{mergeSnpCnv}
\usage{
mergeSnpCnv(.Object, snp.gr = NULL, cnv.gr = NULL)

\S4method{mergeSnpCnv}{RBD}(.Object, snp.gr = NULL, cnv.gr = NULL)
}
\arguments{
\item{snp.gr}{SNP GenomicRanges object}

\item{cnv.gr}{CNV GenomicRanges object}
}
\value{
combined, unique list of genes
}

